+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Better error catching of bad parameter values
 - Catch errors to do with API rate limit
+- [**Stock Time Series Daily node**] - **Bug Fix** Handle non-initialised `msg.payload` in response
+- [**Stock Time Series Intraday node**] - **Bug Fix** Handle non-initialised `msg.payload` in response
 
 ## [0.2.0] - 18/07/2022